chore: default style change
This introduces some changes in the default labstyle.css when created.
It reduced the max-width of the body to 75%, to center the content on
the screen.
It add space between commit message and author, while reducing the one
between author and sha.
It moved the footer text to align to center.